test all three aggregates together
